[arte] Make gltf subfolder, and move Const_Owned to separate header.
