+++ src/lib/arte/arte/gltf/Gltf.h
@@ -1,6 +1,8 @@
 #pragma once
 
 
+#include "Owned.h"
+
 #include <math/Homogeneous.h>
 #include <math/Quaternion.h>
 
@@ -20,52 +22,6 @@ namespace arte {
 namespace gltf {
 
     using EnumType = unsigned int; // this seems to be GLenum type.
-
-    template <class T_indexed>
-    struct Index
-    {
-        using Value_t = std::vector<T_indexed>::size_type;
-
-        Index(Value_t aValue) : 
-            value{aValue}
-        {}
-
-        Index(const Index & aIndex) : 
-            Index{aIndex.value}
-        {}
-
-        Index & operator=(Value_t aValue)
-        {
-            value = aValue;
-            return this;
-        }
-
-        operator Value_t() const
-        { return value; }
-
-        Value_t value;
-    };
-
-    template <class T_indexed>
-    std::ostream & operator<<(std::ostream & aOut, const std::vector<T_indexed> & aIndexVector);
-
-    struct Uri
-    {
-        enum class Type
-        {
-            Data,
-            File,
-        };
-
-        Uri(std::string aString) :
-            string{std::move(aString)}
-        {
-            type = (string.rfind("data:", 0) == 0) ? Type::Data : Type::File;
-        }
-
-        std::string string;
-        Type type;
-    };
 
     struct Accessor;
 
@@ -201,90 +157,6 @@ std::string to_string(gltf::Accessor::ElementType aElementType);
 std::string to_string(gltf::Sampler::Interpolation aInterpolation);
 
 
-// TODO Ad 2022/03/08 Embed the index of the element, to help debug printing.
-template <class T_element>
-class Const_Owned
-{
-    friend class Gltf;
-
-public:
-    Const_Owned(const Gltf & aGltf, const T_element & aElement, gltf::Index<T_element> aIndex);
-
-    operator const T_element &() const
-    { return mElement; }
-
-    const T_element * operator->() const
-    {
-        return &mElement;
-    }
-
-    const T_element & operator*() const
-    {
-        return mElement;
-    }
-
-    gltf::Index<T_element> id() const
-    {
-        return mIndex;
-    }
-
-    template <class T_indexed>
-    Const_Owned<T_indexed> get(gltf::Index<T_indexed> aIndex) const;
-
-    template <class T_indexed>
-    Const_Owned<T_indexed> get(gltf::Index<T_indexed> T_element::* aDataMember) const
-    {
-        return get(mElement.*aDataMember);
-    }
-
-    template <class T_indexed>
-    Const_Owned<T_indexed> get(std::optional<gltf::Index<T_indexed>> T_element::* aDataMember) const
-    {
-        return get(*(mElement.*aDataMember));
-    }
-
-    template <class T_member>
-    std::vector<Const_Owned<T_member>> 
-    iterate(std::vector<gltf::Index<T_member>> T_element::* aMemberIndexVector) const
-    {
-        std::vector<Const_Owned<T_member>> result;
-        for (gltf::Index<T_member> index : mElement.*aMemberIndexVector)
-        {
-            result.emplace_back(mOwningGltf.get(index));
-        }
-        return result;
-    }
-
-    // TODO make lazy iteration
-    template <class T_member>
-    std::vector<Const_Owned<T_member>> iterate(std::vector<T_member> T_element::* aMemberVector) const
-    {
-        std::vector<Const_Owned<T_member>> result;
-        for (std::size_t id = 0; id != (mElement.*aMemberVector).size(); ++id)
-        {
-            result.emplace_back(mOwningGltf, (mElement.*aMemberVector)[id], id);
-        }
-        return result;
-    }
-
-    filesystem::path getFilePath(gltf::Uri T_element::* aMemberUri) const
-    {
-        return mOwningGltf.getPathFor(mElement.*aMemberUri);
-    }
-
-    filesystem::path getFilePath(std::optional<gltf::Uri> T_element::* aMemberUri) const
-    {
-        return mOwningGltf.getPathFor(*(mElement.*aMemberUri));
-    }
-
-
-private:
-    const Gltf & mOwningGltf;
-    const T_element & mElement; 
-    const gltf::Index<T_element> mIndex;
-};
-
-
 class Gltf
 {
 public:
@@ -318,52 +190,6 @@ private:
     std::vector<gltf::BufferView> mBufferViews;
     std::vector<gltf::Accessor> mAccessors;
 };
-
-
-//
-// Implementations
-//
-namespace gltf 
-{
-
-
-    template <class T_indexed>
-    std::ostream & operator<<(std::ostream & aOut, const std::vector<T_indexed> & aIndexVector)
-    {
-        auto begin = std::begin(aIndexVector);
-        auto end = std::end(aIndexVector);
-
-        aOut << "[";
-        if (begin != end)
-        {
-            aOut << *begin;
-            ++begin;
-        }
-        for (; begin != end; ++begin)
-        {
-            aOut << ", " << *begin;
-        }
-        return aOut << "]";
-    }
-
-     
-} // namespace gltf
-
-
-template <class T_element>
-Const_Owned<T_element>::Const_Owned(const Gltf & aGltf, const T_element & aElement, gltf::Index<T_element> aIndex) :
-    mOwningGltf{aGltf},
-    mElement{aElement},
-    mIndex{aIndex}
-{}
-
-
-template <class T_element>
-template <class T_indexed>
-Const_Owned<T_indexed> Const_Owned<T_element>::get(gltf::Index<T_indexed> aIndex) const
-{
-    return mOwningGltf.get(aIndex);
-}
 
 
 } // namespace arte
